### 테스트 코드 수정

Mock.when 시, 패러미터로 any()를 할당 안해서 adoc 문서에 Response 객체가 안 나타나는 현상 해결했어요.

### CourseNotFoundException 위치 이동

advice/exception 폴더로 이동했어요.

+++ src/main/java/org/wooriverygood/api/review/dto/ReviewResponse.java
@@ -2,13 +2,12 @@ package org.wooriverygood.api.review.dto;
 
 import lombok.Builder;
 import lombok.Getter;
-import org.wooriverygood.api.course.dto.CourseResponse;
 import org.wooriverygood.api.review.domain.Review;
 
 @Getter
 public class ReviewResponse {
-    private final int review_id;
-    private final int course_id;
+    private final Long review_id;
+    private final Long course_id;
     private final String review_content;
     private final String review_title;
     private final String instructor_name;
@@ -16,7 +15,7 @@ public class ReviewResponse {
     private final String grade;
 
     @Builder
-    public ReviewResponse(int review_id, int course_id, String review_content, String review_title, String instructor_name, String taken_semyr, String grade) {
+    public ReviewResponse(Long review_id, Long course_id, String review_content, String review_title, String instructor_name, String taken_semyr, String grade) {
         this.review_id = review_id;
         this.course_id = course_id;
         this.review_content = review_content;
