+++ src/main/java/org/wooriverygood/api/review/repository/ReviewRepository.java
@@ -6,5 +6,5 @@ import org.wooriverygood.api.review.domain.Review;
 import java.util.List;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
-    List<Review> findAllByCourseId(int courseId);
+    List<Review> findAllByCourseId(Long courseId);
 }
